forth try
1.str.isEmpty() might cause nullpointException since str might be null .

so I should null check before isEmpty.

2.can make it more clear by using || operator.

3.I need to trim the str before checking whether it's null cause sometimes the user might input a spaces.

String input = "     ";

at that point , input.isEmpty() return false (because it technically has characters-spaces)

But logically we want to treat spaces as empty.

if all characters in this string are space (as defined above), then a String object representing an empty string is returned. since this is the manual of Trim(), so it return's empty string and it trigger isempty method.